Add gitconfig, add Fabric test task, speed up Debian package check

+++ fabfile.py
@@ -1,8 +1,9 @@
 from __future__ import with_statement
 from fabric.api import *
-from fabric.colors import green
+from fabric.colors import green, red
 from fabric.contrib.console import confirm
 from fabric.contrib.files import exists
+from fabric.operations import reboot
 from package_lists import *
 import os
 
@@ -20,6 +21,7 @@ def deploy():
     install_config_files()
     allow_uwsgi_to_control_supervisor()
     set_zsh_as_default_shell()
+    reboot()
 
 def set_hostname():
 	prompt('Enter hostname: ', 'hostname', default='rascal2')
@@ -34,13 +36,22 @@ def install_python_modules():
         print(green('\nInstalling Python module: {0}'.format(module)))
         run('pip install ' + module)
 
+def package_installed(package):
+    cmd = 'dpkg-query -l "{0}" | grep -q ^.i'.format(package)
+    with settings(warn_only=True):
+        result = run(cmd)
+    return result.succeeded
+
 def install_debian_packages():
     print(green('\nUpdating package repository lists'))
     run('apt-get update -y', pty=False)
 
     for package in DEBIAN_PACKAGES_TO_INSTALL:
-        print(green('\nInstalling package: {0}'.format(package)))
-        run('apt-get install -y ' + package, pty=False)
+        if not package_installed(package):
+            print(green('\nInstalling package: {0}'.format(package)))
+            run('apt-get install -y ' + package, pty=False)
+        else:
+            print(green('Package ' + package + ' already installed'))
 
     print(green('\nRemoving unneeded package remnants'))
     run('apt-get autoremove -y', pty=False)
@@ -52,8 +63,11 @@ def disable_bonescript():
 
 def remove_unneeded_debian_packages():
     for package in DEBIAN_PACKAGES_TO_REMOVE:
-        print(green('\nRemoving package: {0}'.format(package)))
-        run('apt-get remove -y ' + package, pty=False)
+        if package_installed(package):
+            print(green('\nRemoving package: {0}'.format(package)))
+            run('apt-get remove -y ' + package, pty=False)
+        else:
+            print(green('Package ' + package + ' already removed.'))
 
 def install_rascal_software():
     print(green('Installing Rascal editor . . .'))
@@ -87,6 +101,7 @@ def install_config_files():
     put('uwsgi.service', '/etc/systemd/system/uwsgi.service')
     run('systemctl enable uwsgi.service')
     put('vimrc', '/root/.vimrc')
+    put('gitconfig', '/root/.gitconfig')
 
 def allow_uwsgi_to_control_supervisor():
     run('echo "chmod=0770 ; socket file mode (default 0700)" >> /etc/supervisor/supervisor.conf')
@@ -99,3 +114,56 @@ def set_zsh_as_default_shell():
     # Enter the new value, or press ENTER for the default
     # Login Shell [/bin/bash]: /bin/zsh
     # curl -L https://github.com/robbyrussell/oh-my-zsh/raw/master/tools/install.sh | sh
+
+### END DEPLOY SECTION ###
+
+### START TEST SECTION ###
+
+@task
+def test():
+    check_debian_packages()
+    check_programs()
+    check_python_modules()
+
+PROGRAMS_TO_TEST = {
+    'nginx': '/usr/sbin/nginx',
+    'uwsgi': '/usr/local/bin/uwsgi',
+    'zsh': '/usr/bin/zsh',
+}
+
+PYTHON_MODULES_TO_TEST_IMPORT = [
+    'gevent',
+    'geventwebsocket',
+    'flask.ext.uwsgi_websocket',
+    'redis',
+    # uwsgi omitted; not a Python module, though installed with pip
+    'pytronics',
+    'bitstring',
+    'webcolors',
+    'flask_login'
+]
+
+def check_debian_packages():
+    for package in sorted(DEBIAN_PACKAGES_TO_INSTALL):
+        if package_installed(package):
+            print(green('Package {0} found.'.format(package)))
+        else:
+            print(red('Package {0} missing!'.format(package)))
+
+def check_programs():
+    for program in sorted(PROGRAMS_TO_TEST.keys()):
+        try:
+            run('which ' + program) == PROGRAMS_TO_TEST[program]
+        except:
+            print(red('Binary for {0} not at {1}.'.format(program, PROGRAMS_TO_TEST[program])))
+        else:
+            print(green('Binary for {0} found at {1}'.format(program, PROGRAMS_TO_TEST[program])))
+
+def check_python_modules():
+    for module in PYTHON_MODULES_TO_TEST_IMPORT:
+        try:
+            run('python -c "__import__(\'' + module + '\')"')
+        except:
+            print(red('Module {0} not installed.'.format(module)))
+        else:
+            print(green('Successfully imported module {0}.'.format(module)))
